Moving to snorkel module naming
