+++ run.sh
@@ -1,10 +1,6 @@
 # Set & move to home directory
-export DDLHOME=$(cd "$(dirname "$0")" && pwd)
-cd "$DDLHOME"
-
-# Make sure home directory is on pythonpath
-export PYTHONPATH="$PYTHONPATH:$DDLHOME"
-export PATH="$PATH:$DDLHOME"
+./set_env.sh
+cd "$SNORKELHOME"
 
 # Make sure parser is installed
 PARSER="parser/stanford-corenlp-3.6.0.jar"
