+++ snorkel/parser.py
