+++ test/data/CDR_TestSet_docs.pkl
@@ -2,7 +2,7 @@
 ccopy_reg
 _reconstructor
 p2
-(cddlite_parser
+(csnorkel.parser
 Document
 p3
 c__builtin__
