+++ tutorial/load_dictionaries.py
@@ -1,5 +1,8 @@
 import bz2
 import csv
+import os
+
+ROOT = os.path.join(os.environ['SNORKELHOME'], 'tutorial/data/dicts/')
 
 def load_bioportal_csv_dictionary(filename):
     """Load BioPortal Ontologies--http://bioportal.bioontology.org/"""
@@ -20,28 +23,28 @@ def load_disease_dictionary():
     d = set()
       
     # UMLS SemGroup Disorders
-    dictfile = "data/dicts/umls_disorders_v2.bz2"
+    dictfile = ROOT + "/umls_disorders_v2.bz2"
     diseases = [line.strip().split("\t")[0] for line in bz2.BZ2File(dictfile, 'rb').readlines()]
     d.update(w for w in diseases if not w.isupper())
 
     # Orphanet Rare Disease Ontology
-    d.update(w for w in load_bioportal_csv_dictionary("data/dicts/ordo.csv") if not w.isupper())
+    d.update(w for w in load_bioportal_csv_dictionary(ROOT + "/ordo.csv") if not w.isupper())
     
     # Human Disease Ontology 
-    d.update(w for w in load_bioportal_csv_dictionary("data/dicts/DOID.csv") if not w.isupper())
+    d.update(w for w in load_bioportal_csv_dictionary(ROOT + "/DOID.csv") if not w.isupper())
       
     # ------------------------------------------------------------
     # remove cell dysfunction terms
-    dictfile = "data/dicts/cell_molecular_dysfunction.txt"
+    dictfile = ROOT + "/cell_molecular_dysfunction.txt"
     remove_terms = set(line.strip().split("\t")[0] for line in open(dictfile).readlines())
 
     # remove geographic areas terms
-    dictfile = "data/dicts/umls_geographic_areas.txt"
+    dictfile = ROOT + "/umls_geographic_areas.txt"
     remove_terms.update(line.strip().split("\t")[0] for line in open(dictfile).readlines())
     d = d.difference(remove_terms)
     
     # remove stopwords
-    dictfile = "data/dicts/stopwords.txt"
+    dictfile = ROOT + "/stopwords.txt"
     stopwords = [line.strip().split("\t")[0] for line in open(dictfile).readlines()]
     d = [w for w in list(d) if w.lower() not in stopwords]
     return d
@@ -54,15 +57,15 @@ def load_acronym_dictionary():
     a = set()
     
     # UMLS disorders
-    dictfile = "data/dicts/umls_disorders_v2.bz2"
+    dictfile = ROOT + "/umls_disorders_v2.bz2"
     diseases = [line.strip().split("\t")[0] for line in bz2.BZ2File(dictfile, 'rb').readlines()]
     a.update(w for w in diseases if w.isupper())
     
     # Orphanet Rare Disease Ontology
-    a.update(w for w in load_bioportal_csv_dictionary("data/dicts/ordo.csv") if w.isupper())
+    a.update(w for w in load_bioportal_csv_dictionary(ROOT + "/ordo.csv") if w.isupper())
     
     # Human Disease Ontology 
-    a.update(w for w in load_bioportal_csv_dictionary("data/dicts/DOID.csv") if w.isupper())
+    a.update(w for w in load_bioportal_csv_dictionary(ROOT + "/DOID.csv") if w.isupper())
     
     # filter by char length
     a = [w for w in list(a) if len(w) > 1]
